用户ID设置Long #13

+++ seen-common/seen-api/src/main/java/com/songchi/seen/account/service/CoinAccountBalanceService.java
@@ -10,9 +10,9 @@ import java.util.Set;
  * @since 2023-01-01
  */
 public interface CoinAccountBalanceService {
-    Map<Integer, Integer> accountIdCoinAmount(Set<Integer> accountIds);
+    Map<Long, Integer> accountIdCoinAmount(Set<Long> accountIds);
 
-    boolean set(Integer accountId, Integer amount);
+    boolean set(Long accountId, Integer amount);
 
-    boolean add(Integer accountId, Integer offsetAmount);
+    boolean add(Long accountId, Integer offsetAmount);
 }
