+++ seen-common/seen-api/src/main/java/com/songchi/seen/trade/service/TradeService.java
@@ -9,12 +9,12 @@ import com.songchi.seen.coin.enumeration.TradeType;
  * @since 2023-03-11
  */
 public interface TradeService {
-    Integer trade(Integer inAccountId, Integer outAccountId, Integer coinAmount, TradeType tradeType,
+    Integer trade(Long inAccountId, Long outAccountId, Integer coinAmount, TradeType tradeType,
                   String description);
 
-    Integer sysInTrade(Integer inSysAccountId, Integer outAccountId, Integer coinAmount, TradeType tradeType, String description);
+    Integer sysInTrade(Long inSysAccountId, Long outAccountId, Integer coinAmount, TradeType tradeType, String description);
 
-    Integer sysOutTrade(Integer inAccountId, Integer outSysAccountId, Integer coinAmount, TradeType tradeType, String description);
+    Integer sysOutTrade(Long inAccountId, Long outSysAccountId, Integer coinAmount, TradeType tradeType, String description);
 
-    Integer sysTrade(Integer inSysAccountId, Integer outSysAccountId, Integer coinAmount, TradeType tradeType, String description);
+    Integer sysTrade(Long inSysAccountId, Long outSysAccountId, Integer coinAmount, TradeType tradeType, String description);
 }
