+++ seen-micro-database/seen-chat/src/main/java/com/songchi/seen/chat/po/ChatUserPO.java
@@ -27,11 +27,11 @@ public class ChatUserPO {
     /**
      * 用户ID
      */
-    private Integer userId;
+    private Long userId;
     /**
      * 好友ID
      */
-    private Integer friendUserId;
+    private Long friendUserId;
     /**
      * 更新时间
      */
